+++ src/client/java/fr/madu59/obe/client/util/blockentity/ChestUtil.java
@@ -22,21 +22,21 @@ import net.minecraft.world.level.block.state.properties.ChestType;
 
 public class ChestUtil {
     private static final Calendar calendar = Calendar.getInstance();
-    private static final boolean xmasTexture = calendar.get(2) + 1 == 12 && calendar.get(5) >= 24 && calendar.get(5) <= 26;
+    public static final boolean isXmas = calendar.get(2) + 1 == 12 && calendar.get(5) >= 24 && calendar.get(5) <= 26;
 
     public static ResourceLocation getMaterial(BlockState state) {
         Block block = state.getBlock();
         ChestType chestType = BackportUtil.getValueOrElse(state, ChestBlock.TYPE, ChestType.SINGLE);
         if (block instanceof EnderChestBlock) {
             return Sheets.ENDER_CHEST_LOCATION.texture();
-        } else if (xmasTexture) {
+        } else if (isXmas) {
             return chooseMaterial(chestType, Sheets.CHEST_XMAS_LOCATION, Sheets.CHEST_XMAS_LOCATION_LEFT, Sheets.CHEST_XMAS_LOCATION_RIGHT);
         } else {
             return block instanceof TrappedChestBlock ? chooseMaterial(chestType, Sheets.CHEST_TRAP_LOCATION, Sheets.CHEST_TRAP_LOCATION_LEFT, Sheets.CHEST_TRAP_LOCATION_RIGHT) : chooseMaterial(chestType, Sheets.CHEST_LOCATION, Sheets.CHEST_LOCATION_LEFT, Sheets.CHEST_LOCATION_RIGHT);
         }
     }
 
-    private static ResourceLocation chooseMaterial(ChestType chestType, Material material, Material material2, Material material3) {
+    public static ResourceLocation chooseMaterial(ChestType chestType, Material material, Material material2, Material material3) {
         if(chestType == null) return material.texture();
         return switch (chestType) {
             case RIGHT -> material3.texture();
@@ -65,7 +65,7 @@ public class ChestUtil {
         poseStack.translate(-0.5F, -0.5F, -0.5F);
     }
 
-    public static void transform(BlockState state, PoseStack poseStack, BlockEntity entity){
+    public static void transform(BlockState state, BlockEntity entity, PoseStack poseStack){
         transform(state, poseStack);
     }
 
@@ -97,7 +97,7 @@ public class ChestUtil {
     }
 
     @Deprecated
-    public static void transformChest(BlockState state, PoseStack poseStack, BlockEntity be){
-        transform(state, poseStack, be);
+    public static void transformChest(BlockState state, BlockEntity be, PoseStack poseStack){
+        transform(state, be, poseStack);
     }
 }