+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # What's new:
 - Bug fixes
 - Added compat with all chests using the Blueprint mod (Woodworks and more)
+
+# Patch 1:
+- Added compatiblity with Every Compat
+- Added compatibility with Carved Wood